write some code for change dimension

+++ common/src/main/kotlin/g_mungus/vlib/api/VLibGameUtils.kt
@@ -1,13 +1,17 @@
 package g_mungus.vlib.api
 
 import g_mungus.vlib.VLib.LOGGER
+import g_mungus.vlib.structure.StructureManager.enqueueTemplateForAssembly
 import net.minecraft.core.BlockPos
 import net.minecraft.core.Vec3i
+import net.minecraft.resources.ResourceKey
 import net.minecraft.resources.ResourceLocation
 import net.minecraft.server.level.ServerLevel
+import net.minecraft.world.level.Level
 import net.minecraft.world.level.block.Blocks
 import net.minecraft.world.level.levelgen.structure.templatesystem.StructureTemplate
 import net.minecraft.world.level.levelgen.structure.templatesystem.StructureTemplateManager
+import org.joml.Vector3d
 import org.joml.primitives.AABBic
 import org.valkyrienskies.core.api.ships.ServerShip
 import org.valkyrienskies.core.api.ships.Ship
@@ -20,7 +24,7 @@ import kotlin.random.Random
  * @param structurePath The namespace or directory to save the structure template. Can either be in the format <namespace> or <namespace:folder/>"
  * @param level The relevant ServerLevel
  * @param blockPos A block position in the shipyard on the ship to be saved
- * @param withEntities Whether to save non-player entities from the ship. I don't think it works right now.
+ * @param withEntities Whether to save non-player entities from the ship. I haven't tested it yet.
  * @param deleteAfter Whether to delete the ship after it has been saved to a structure template.
  *
  * @return nothing
@@ -40,7 +44,7 @@ fun saveShipToTemplate(structurePath: String, level: ServerLevel, blockPos: Bloc
  * @param structurePath The namespace or directory to save the structure template. Can either be in the format <namespace> or <namespace:folder/>"
  * @param level The relevant ServerLevel
  * @param shipId The shipID of the ship to be saved
- * @param withEntities Whether to save non-player entities from the ship. I don't think it works right now.
+ * @param withEntities Whether to save non-player entities from the ship. I haven't tested it yet.
  * @param deleteAfter Whether to delete the ship after it has been saved to a structure template.
  *
  * @return nothing
@@ -53,14 +57,9 @@ fun saveShipToTemplate(structurePath: String, level: ServerLevel, shipId: Long, 
     }
 
     val structureTemplateManager = level.structureManager
-
-
     structureTemplateManager.save(getStructureTemplate(structurePath, level, ship, withEntities, structureTemplateManager).second)
 
-    if (deleteAfter) {
-        level.shipObjectWorld.deleteShip(ship as ServerShip)
-    }
-
+    if (deleteAfter) { level.shipObjectWorld.deleteShip(ship as ServerShip) }
 }
 
 private fun getStructureTemplate (structurePath: String, level: ServerLevel, ship: Ship, withEntities: Boolean, structureTemplateManager: StructureTemplateManager): Pair<StructureTemplate, ResourceLocation> {
@@ -102,5 +101,42 @@ private fun getSize(aabb: AABBic?): Vec3i {
     } else {
         throw Exception("Why doesn't the ship have a shipyard AABB?")
     }
+}
+
+
+/**
+ * Untested
+ */
+fun changeDimension(ship: Ship, serverLevel: ServerLevel, targetDimension: ResourceKey<Level>) {
+    val structureTemplateManager = serverLevel.structureManager
+    val structureTemplate = getStructureTemplate(
+        structurePath = "vlib:interdimensional/",
+        level = serverLevel,
+        ship = ship,
+        withEntities = true,
+        structureTemplateManager = structureTemplateManager
+    )
+
+    val targetLevel = serverLevel.server.getLevel(targetDimension)?: return
+
+    val shipPos = ship.shipAABB?.center(Vector3d())
+    var shipDest = BlockPos(0, 0, 0)
+
+    if (shipPos != null) {
+        if (targetLevel.maxBuildHeight > (shipPos.y + ship.shipAABB!!.maxY() - ship.shipAABB!!.minY())) {
+            shipDest = BlockPos(shipPos.x.toInt(), shipPos.y.toInt(), shipPos.z.toInt())
+        } else {
+            shipDest = BlockPos(shipPos.x.toInt(), targetLevel.maxBuildHeight - ship.shipAABB!!.maxY() + ship.shipAABB!!.minY(), shipPos.z.toInt())
+        }
+    }
+
+    enqueueTemplateForAssembly(
+        structureTemplate = structureTemplate.first,
+        serverLevel = targetLevel,
+        shipDest
+    )
+
+    serverLevel.shipObjectWorld.deleteShip(ship as ServerShip)
+
 }
